Update README.md
Chatbot- An application to make a easy and descriptive data's in a very brief manner so that every details can make a change out. This can help to manage the question and answers in a very good manner. Every management has it's own details so the calculable report is to have a chatbot like this.